ui/gtk3: Enumerate items of a list
The items within a list are finally being enumerated when the user
clicks on the name of the list. Deleting a TODO item is also possible
by clicking on the lil X icon of a row (same as with lists).

The editing feature was temporarily disabled because the click on the
list name is now leading the user to the items view. Not sure about
what's the best UI for that.

Things to watch out:

 * There's a lot of duplicated code between the lists & items
   views. The only things made [cumbersomely] reusable for now were
   the `NewItemDialog` and the `Toolbar`. This can certainly be
   improved later.

 * The communication between the UI elements is very rudimentary, the
   instance of the `MainPanel` class has been passed to most of the UI
   components. A standard needs to be chosen.