ui/gtk3: Different icons & back button in the toolbar
The back button shows up when the constructor of the toolbar receives
a third parameter that is a boolean. It's not the best way to do that
but that might be the last change on the toolbar for a while so it
didn't get complex enough to deserve a more interesting solution.